requirements.txt
groq>=0.11.0
python-dotenv>=1.0.1
ddgs>=2.0.0
arxiv>=2.1.3
python-dotenv==1.2.1
requests>=2.31.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
